backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.23
asyncpg>=0.29.0
alembic>=1.12.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
boto3>=1.34.0
botocore>=1.34.0
python-jose[cryptography]>=3.3.0
python-dateutil>=2.8.2
clerk-backend-api>=3.0.5
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.28.1
ruff>=0.1.6
mypy>=1.7.1